requirements.txt
tqdm==4.60.0
ray==1.3.0
scipy==1.6.2
pytorch-lightning==1.2.6
numpy==1.19.2
matplotlib==3.3.4
scikit-image==0.17.2
scikit-learn==0.24.1
test-tube==0.7.5
cupy-cuda102==8.6.0
pytorch-fid==0.2.0